server/pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.73.1",
    "livekit>=1.1.5",
    "livekit-agents[google,images,openai]>=1.5.4",
    "livekit-plugins-bey>=1.5.4",
    "livekit-plugins-google>=1.5.4",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "livekit-plugins-silero>=1.5.4",
    "livekit-plugins-turn-detector>=1.5.4",
    "livekit-protocol>=1.1.5",
    "python-dotenv>=1.2.2",
    "tzdata>=2026.1",
]
